exercice 10,11,12 terminé, exo 7 refait avec boolean

+++ src/exo9AfficherLeMoisCorrespondant.java
@@ -3,7 +3,8 @@ public class exo9AfficherLeMoisCorrespondant {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.print("Afficher le numero du mois : ");
-        int mois = input.nextInt();
+        byte mois = input.nextByte();
+       // uniquement test d'égalité on fait switch case, attention ne pas oublié break
         switch (mois)
         {
             case 1:
